+++ src/extensions/rules/rulesext.h
@@ -282,8 +282,13 @@ public:
      */
     bool IsUseBridgeHealth;
 
-	/*
-    * The armor type used by bridges for damage calculation. Only used when Bridge Health mechanism is enabled.
-    */
+	/**
+     * The armor type used by bridges for damage calculation. Only used when Bridge Health mechanism is enabled.
+     */
     ArmorType BridgeArmor;
+
+	/**     
+	 *  Determines whether the cloaked technos can trigger cell tags when used with Entered By trigger event.
+     */
+    bool IsCellTagsIgnoreStealth;
 };